Fix template match timing tracking bugs
Core issue: MatchTemplatePanel never called StartTracking(), causing elapsed times to show as Unix timestamps instead of seconds.

Changes:
- Add StartTracking() call in MatchTemplatePanel.cpp:1492 (matches FindCTFPanel.cpp:894 pattern)
- Fix database schema encoding for ELAPSED_TIME_SECONDS ('r'→'l' in database_schema.h:61)
- Change elapsed_time_seconds type from double to long throughout codebase
- Use .ToLong() instead of .ToDouble() to avoid wxLongLong conversion bugs
- Add missing template_match_id assignment in database.cpp:2050
- Show newest results first (ORDER BY TEMPLATE_MATCH_ID DESC) for better UX
- Update CLAUDE.md with debugging lessons learned

+++ src/core/template_matching.h
@@ -28,7 +28,7 @@ class TemplateMatchJobResults {
     long     job_id; // This is the search_id in the database
     long     template_match_id; // The TEMPLATE_MATCH_ID for this specific result
     long     datetime_of_run;
-    double   elapsed_time_seconds; // Time elapsed from start to completion
+    long     elapsed_time_seconds; // Time elapsed from start to completion (in seconds)
     long     image_asset_id;
     long     ref_volume_asset_id;
     wxString symmetry;
